+++ MekHQ/src/mekhq/campaign/personnel/lifeEvents/NewYearsDayAnnouncement.java
@@ -41,7 +41,7 @@ import java.util.List;
 import megamek.common.annotations.Nullable;
 import megamek.logging.MMLogger;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.personnel.Person;
 import mekhq.gui.baseComponents.immersiveDialogs.ImmersiveDialogSimple;
