+++ MekHQ/src/mekhq/campaign/universe/Planet.java
@@ -42,11 +42,11 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.util.StdConverter;
 import megamek.codeUtilities.ObjectUtility;
 import megamek.common.ITechnology.TechRating;
-import megamek.common.annotations.Nullable;
 import megamek.common.TargetRoll;
+import megamek.common.annotations.Nullable;
 import megamek.common.universe.FactionTag;
 import megamek.logging.MMLogger;
-import mekhq.campaign.CampaignOptions;
+import mekhq.campaign.campaignOptions.CampaignOptions;
 import mekhq.campaign.universe.PlanetarySystem.PlanetaryRating;
 import mekhq.campaign.universe.PlanetarySystem.PlanetarySophistication;
 import mekhq.campaign.universe.enums.HPGRating;
